+++ bocadillo/redirection.py
@@ -4,9 +4,14 @@ from starlette.responses import RedirectResponse
 class Redirection(Exception):
     """Raised when redirecting to another page."""
 
-    def __init__(self, url: str):
-        self._response = RedirectResponse(url=url)
+    def __init__(self, url: str, permanent: bool):
+        self._url = url
+        self._permanent = permanent
+
+    @property
+    def status_code(self) -> int:
+        return 301 if self._permanent else 302
 
     @property
     def response(self) -> RedirectResponse:
-        return self._response
+        return RedirectResponse(url=self._url, status_code=self.status_code)